Simplify code for instantiating jobs
- Split Job class into a "Job" interface and a "AbstractJob" class
  that PersistentJob and VolatileJob are based on.
- Remove the Job.JobBuilder class. It was very convoluted and not
  needed, and a bit confusing because JobManager.JobBuilder also
  exists.
- Delete RuntimeConfigurator class. Now it is more clear that the job
  implementations (PersistentJob and VolatileJob) themselves are
  responsible for setting the event bus and job monitor.
