+++ framework-persistence/src/test/java/org/daisy/pipeline/persistence/impl/job/PersistentJobTest.java
@@ -5,12 +5,8 @@ import java.util.List;
 
 import org.daisy.common.priority.Priority;
 import org.daisy.pipeline.job.Job;
-import org.daisy.pipeline.job.Job.JobBuilder;
 import org.daisy.pipeline.job.JobId;
 import org.daisy.pipeline.persistence.impl.Database;
-import org.daisy.pipeline.persistence.impl.job.PersistentJob;
-import org.daisy.pipeline.persistence.impl.job.PersistentJobContext;
-import org.daisy.pipeline.persistence.impl.job.PersistentJob.PersistentJobBuilder;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -30,12 +26,10 @@ public class PersistentJobTest   {
 		db=DatabaseProvider.getDatabase();
 		System.setProperty("org.daisy.pipeline.iobase",System.getProperty("java.io.tmpdir"));
 		PersistentJobContext.setScriptRegistry(new Mocks.DummyScriptService(Mocks.buildScript()));
-		JobBuilder builder= new PersistentJobBuilder(db).withContext(Mocks.buildContext());
-		job =(PersistentJob) builder.build();//new PersistentJob(Job.newJob(Mocks.buildContext()),db);
+		job = new PersistentJob(db, Mocks.buildContext());
 		id=job.getContext().getId();
-                //high priority
-                builder= new PersistentJobBuilder(db).withContext(Mocks.buildContext());
-		jobHigh =(PersistentJob) builder.withPriority(Priority.HIGH).build();//new PersistentJob(Job.newJob(Mocks.buildContext()),db);
+		// high priority
+		jobHigh = new PersistentJob(db, Mocks.buildContext(), Priority.HIGH);
 		idHigh=jobHigh.getContext().getId();
 	}
 	@After
